コメント追加: カテゴリ更新処理を説明
