+++ docs/architecture_proposals_ja.md
@@ -1,0 +1,21 @@
+# ウィジェット肥大化への対応案
+
+本アプリではクリーンアーキテクチャを採用していますが、画面ごとのウィジェットが大きくなりがちです。一般的なアーキテクチャの考え方に沿った対策案を以下にまとめます。
+
+## 1. ウィジェットの分割
+- 画面内のコンポーネントを小さな StatelessWidget / StatefulWidget として切り出し、`lib/widgets` 配下に配置します。
+- これによりファイル単位で役割が明確になり、テストもしやすくなります。
+- 例として `InventoryCard` を `lib/widgets/inventory_card.dart` へ分離しました。
+
+## 2. 状態管理の導入
+- UI とビジネスロジックを分離するため、Riverpod や BLoC などの状態管理ライブラリを利用します。
+- 画面側は状態の監視とイベント通知のみに集中させ、データ取得や加工は ViewModel / Bloc で行います。
+
+## 3. MVVM などのパターン
+- Clean Architecture をベースに、プレゼンテーション層では MVVM や Presenter パターンを採用する方法もあります。
+- ViewModel が UseCase を呼び出し、画面は ViewModel を監視するだけにすると見通しが良くなります。
+
+## 4. ルーティングの整理
+- 画面遷移が複雑になった場合は、Navigator 2.0 (Router API) を利用してルーティングを一元管理します。
+
+以上の方法を組み合わせることで、ウィジェットの肥大化を防ぎ、保守性を高めることができます。
