+++ test/inventory_card_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:oouchi_stock/main.dart';
+import 'package:oouchi_stock/widgets/inventory_card.dart';
 import 'package:oouchi_stock/domain/entities/inventory.dart';
 
 void main() {
